separate payload stitching into separate module
